feature/create endpoint for search rides
