+++ apps/ride_manager/apis/__init__.py
@@ -1,8 +1,8 @@
 from .affected_place_viewset import AffectedPlaceViewSet
 from .person_register_viewset import PersonRegisterViewSet
+from .ride_search_viewset import RideSearchViewSet
 from .ride_viewset import RideViewset
 from .validate_phone_check_code import ValidatePhoneCheckCodeApiView
 from .validate_phone_send_code import ValidatePhoneSendCodeApiView
 from .vehicle_viewset import VehicleViewset
 from .voluntary_viewset import VoluntaryViewset
-from .ride_search_viewset import RideSearchViewSet
